tests: inline more steps
