+++ spec/features/comments/commenting_spec.rb
@@ -3,6 +3,10 @@
 require 'rails_helper'
 
 feature "Commenting" do
+  def i_write_a_new_comment body
+    first("#comment_body").set body
+  end
+
   background do
     i_log_in_as_a_catalog_editor_named "Batiatus"
     create :feedback
