+++ spec/features/markdown/autocompletion_spec.rb
@@ -3,6 +3,11 @@
 require 'rails_helper'
 
 feature "Markdown autocompletion", as: :editor, js: true do
+  def i_am_on_a_page_with_a_textarea_with_markdown_preview_and_autocompletion
+    i_go_to 'the open issues page'
+    i_follow "New"
+  end
+
   def markdown_textarea
     find(".preview-area textarea")
   end
