+++ spec/features/markdown/preview_spec.rb
@@ -5,8 +5,9 @@ require 'rails_helper'
 feature "Preview markdown", as: :editor, skip_ci: true, js: true do
   scenario "Previewing references markdown" do
     giovanni_1809 = create :any_reference, author_string: "Giovanni, S. ", year: 1809
-    i_am_on_a_page_with_a_textarea_with_markdown_preview_and_autocompletion
 
+    i_go_to 'the open issues page'
+    i_follow "New"
     fill_in "issue_description", with: "See: #{Taxt.ref(giovanni_1809.id)}"
     click_button "Rerender preview"
     i_should_see "See: Giovanni, 1809"
